Move LegacyConfigTransformer to connection_spec file
also add tests
